requirements.txt
click
python-gnupg
git+git://github.com/ammgws/google_auth.git
git+git://github.com/ammgws/hangouts_client.git
